+++ planning/behavior_path_avoidance_by_lane_change_module/include/behavior_path_avoidance_by_lane_change_module/interface.hpp
@@ -17,7 +17,7 @@
 
 #include "behavior_path_avoidance_by_lane_change_module/data_structs.hpp"
 #include "behavior_path_avoidance_by_lane_change_module/scene.hpp"
-#include "behavior_path_planner/scene_module/lane_change/interface.hpp"
+#include "behavior_path_lane_change_module/interface.hpp"
 
 #include <rclcpp/rclcpp.hpp>
