+++ planning/behavior_path_avoidance_by_lane_change_module/package.xml
@@ -19,6 +19,7 @@
   <buildtool_depend>eigen3_cmake_module</buildtool_depend>
 
   <depend>behavior_path_avoidance_module</depend>
+  <depend>behavior_path_lane_change_module</depend>
   <depend>behavior_path_planner</depend>
   <depend>behavior_path_planner_common</depend>
   <depend>motion_utils</depend>
