+++ planning/behavior_path_external_request_lane_change_module/test/test_behavior_path_planner_node_interface.cpp
@@ -49,8 +49,6 @@ std::shared_ptr<BehaviorPathPlannerNode> generateNode()
     ament_index_cpp::get_package_share_directory("behavior_path_planner");
 
   std::vector<std::string> module_names;
-  module_names.emplace_back("behavior_path_planner::LaneChangeRightModuleManager");
-  module_names.emplace_back("behavior_path_planner::LaneChangeLeftModuleManager");
   module_names.emplace_back("behavior_path_planner::ExternalRequestLaneChangeRightModuleManager");
   module_names.emplace_back("behavior_path_planner::ExternalRequestLaneChangeLeftModuleManager");
